# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sytemap API Documentation",
        "description": "API documentation for Sytemap application with OTP verification and user management",
        "contact": {
            "email": "admin@sytemap.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/v1/admin/referrals": {
            "get": {
                "tags": [
                    "Admin - Referrals"
                ],
                "summary": "Get list of all agents and their referrals",
                "description": "This endpoint returns all agents with their referred users. Only accessible by admin.",
                "operationId": "4a2ca1962e6f72861575ac3152915c56",
                "responses": {
                    "200": {
                        "description": "List of agents and their referrals",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "agent_id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "agent_name": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "agent_email": {
                                                        "type": "string",
                                                        "example": "john@example.com"
                                                    },
                                                    "referrals_count": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "referred_users": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 12
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "Mary Smith"
                                                                },
                                                                "email": {
                                                                    "type": "string",
                                                                    "example": "mary@example.com"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}